backend/services/gemini_live_proxy/requirements.txt
websockets>=16.0
google-auth>=2.38.0
google-auth[requests]>=2.38.0
requests>=2.32.0
certifi>=2025.11.10
python-dotenv>=1.2.2
